Break out playlist selector logic Pull the basic playlist selector into a separate file so there's a shared place to build alternates.
